Modified app.py to create db tables too
This is instead of just run.py creating the tables and relying
on the WSGI app being ran by uWSGI or Heroku.

It only creates the db tables if the app.config['DEBUG'] (also
on app.py) is set to True. If it is set to False, it will only
create db tables when executing run.py.